core: Fix logfile permissions, attila style
In the case where we're started as root, there are issues in the
configuration and we have no logfile, such as when a user first
installs Nagios and fires it up via the init-script through a
reboot, we will fail to open the logfile after having dropped
root privilegs.

With this patch, we chown the logfile and debug logfile to the
destination user just before we call setgid() and setuid(),
which is the hardcore way of making sure we can always write
to our logfile.

In the case where we're running as a non-root user, the call
is never made and we rely on the check to see if we can open
the logfile after privileges are dropped for proper logging.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2618 5f96b256-904b-4d8d-8c98-d829582c6739