downtime: Fix crash when using distributable downtime commands
The function that implements the distributable downtime commands kept a
pointer to the next downtime, which is a rubbish idea when the
unschedule_downtime deletes any triggered_by downtimes, and scheduled
downtimes triggered by this one typically start at the first element
right after this one.

To get nagios to crash using the old code, get a configuration with a
few hosts, make all of them depend on one single host, schedule and
propagate a triggered host downtime on the parent, and then unschedule
the downtime on the parent using, for example,
DEL_DOWNTIME_BY_HOST_NAME.

This prevents the crash by first locating any matching downtimes and
copying the information about the downtime type and id to a separate
downtime list, and only then try to unschedule anything. This might lead
to unschedule_downtime having to look for downtimes that no longer
exist, which is slow, but it should only really happen for
DEL_DOWNTIME_BY_START_TIME_COMMENT. I can see no easy way around this
without breaking API and/or ABI.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2620 5f96b256-904b-4d8d-8c98-d829582c6739