runcmd: Fix quoting issues when running commands
The important part of this patch is one line that makes sure that if
we're in single quoted mode, we should ignore '\' as an escape
character.

Without this, foo\\bar was interprented exactly the same as 'foo\\bar'.
With it, the former is executed as one backslash, the latter as two.

There's is also a new test-runcmd, with a small set of test cases for
the command quoter.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2627 5f96b256-904b-4d8d-8c98-d829582c6739