checks: Check command output does not show up pre-escaped
Why would you think it did?

Since we assumed that check output was escaped, we tried to unescape
it, and that becomes highly weird.

So, this bug meant that, unless a command is quoted (which, as shown in
the parent commit, you couldn't do), if you're trying to test quoting by
writing a check command that echoes your input, you needed four
backslashes in your input to get one in your output. Trying to read the
quad-escaped commands in the input, in turn, quickly makes a user want
to escape from this application.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2628 5f96b256-904b-4d8d-8c98-d829582c6739