lib/worker: Add spawn_named_helper()
Setting argv[0] to some arbitrary name causes that arbitrary name
to show up as the process "title" (same as name, really) in most
programs, such as pidof, ps and top.

We don't use that feature yet, but when we get asked to spawn
particular workers (which will most likely happen "any day now")
it's nifty to be able to separate the workers from each other so
while also grouping them together. This patch makes that possible
without changing the current behaviour at all.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2630 5f96b256-904b-4d8d-8c98-d829582c6739