lib/worker: Do one last round of output gathering in finish_job()
But only if the filedescriptors aren't -1 ofcourse. This works
around a bug (or at least an oddity) with the epoll() interface
on Linux kernels older than 2.6.31. Since later kernels will
do the extra calls very rarely anyway (when a plugin times out
and we kill it ourselves), we make them unconditional to make
sure the code gets some exercise.

To avoid unpleasantries of re-entering parent function and things
like that, we revamp the static functions gather_output() and
finish_job() somewhat.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2651 5f96b256-904b-4d8d-8c98-d829582c6739