xodtemplate: Fix same-host service dependencies with inexact master
Weird subject, but here's the thing. In Nagios 3 we supported
same-host servicedependencies where the host was defined by having
exact settings for the dependent services (such as a servicegroup
or a combination of dependent_hostgroup_name/dependent_host_name
AND dependent_service_description). With Nagios 4-ish, that test
slipped through the testing net, even though we actually have a
test configuration just like that checked in alongside the code.

With this patch, we no longer bomb out when given servicedependencies
such as this:
  define servicedependency {
     service_description           some-service
     dependent_servicegroup        some-service-group
     notification_failure_criteria w,c
  }

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2659 5f96b256-904b-4d8d-8c98-d829582c6739