common/objects: Make qsort() comparison functions a lot faster
Since host and service objects are sorted lexicographically and
carry an unsigned int as id, which is incremental based on the
sort order, we can just compare integers rather than run full
string comparisons when sorting host and service dependencies.

I'm unsure how much impact this has on config reload times, but
it will surely make it slightly faster. The fact that we remove
more code than we add makes that a double win.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2660 5f96b256-904b-4d8d-8c98-d829582c6739