Corrected comparison operator in service freshness check logic.
The service freshness check logic was resetting the expiration after every
short (relative to the freshness threshold) Nagios outage for service 
checks but preserving the expiration if Nagios had been down for a long 
time in comparison with the freshness threshold.

This logic was opposite the host freshness check logic, which has been correct
for quite some time.


git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2668 5f96b256-904b-4d8d-8c98-d829582c6739