test: Remove ridonculous tests
The 620history.t and 621 extinfo.t tests only provide a html lint
scan, but relies on Test::WWW::Mechanize::CGI, which no current
distribution ships by default. Since that means it's almost certainly
too much hassle for package developers to bother with, we get rid
of them completely so we have a chance of getting people (and me) to
run 'make test' on a regular basis, since it does catch actual
errors from time to time and marks incompatibilities when they're
introduced.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2670 5f96b256-904b-4d8d-8c98-d829582c6739