Let cgi's use object pointers too
It's costly and inefficient to use two separate object layouts
depending on whether we're building for cgi's or the daemon, and
especially so when the separate object layouts means we require a
lot of extra code to provide the same functionality to both parts,
except one codepath is slower than the other.

With this patch we streamline the objects and let the cgi's use
object pointers the same way the core does when looking up members
and whatnot.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2673 5f96b256-904b-4d8d-8c98-d829582c6739