lib/nsutils: Welcome to the world
This breaks apart the current lib/worker.c and removes some of the
not-really-compatibility macros from lib/lnag-utils.h and puts them
in a file of their own, so we can implement a sane naming standard
for the functions and macros that actually relate to worker
communications.

While at it, we give the newly-homed functions new names, with their
old names as aliases, and provide deprecation notices for them so
our module authors know they'll be gone in 4.1.0.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2676 5f96b256-904b-4d8d-8c98-d829582c6739