lib/worker: Rename send_kvvec() to worker_send_kvvec()
Getting and using some sort of namespace for the worker code is
nifty. This is at least a step in the right direction. Since the
library is intertwined with the default core worker, that's gonna
be pretty hard, so we'll simply have to take what we can and run
with it.

The patch is bigger than it should have been, since we have to
keep send_kvvec() in both the library and the headers until it's
had its grace deprecation period.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2684 5f96b256-904b-4d8d-8c98-d829582c6739