lib: Make NAGIOS_DEPRECATED work for compilers older than gcc 4.5.0
The optional message argument to the deprecated attribute was added
in gcc 4.5.0, so older versions of gcc fail to build Nagios without
this patch.

As a neat side-effect of how we implement the thing, we also get a
NAGIOS_MKVERSION() macro, which can be used to set API versions for
various parts of Nagios and libnagios, which will come in handy for
handling incompatibilities.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2692 5f96b256-904b-4d8d-8c98-d829582c6739