Macros: fix of invalid free for TOTALHOSTSERVICES* macros
The $TOTALHOSTSERVICES*$ macros were computed all at the same time, with the
result stashed in mac->x using the mkstr method (from worker.c? why worker.c?)

The mkstr uses an static internal buffer of usally 256 strings, of each 32
bytes, for generation of strings, used in macros. When computing
TOTALHOSTSERVICES*, the valuse is, as said, stashed in mac->x[], which is freed
 at cleanup, with some exceptions, and this isn't one of them.

Those needs to be exceptions, so static function variables isn't freed

Signed-off-by: Max Sikström <msikstrom@op5.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2693 5f96b256-904b-4d8d-8c98-d829582c6739