libnagios: Add the 'fanout' api
Fanouts are useful to hold expiring 'long'-indexed data where
the span between smallest and largest key can be too large and
change too often for it to be practical to maintain a "real" array.

For Nagios, that currently matches downtime, comments, and jobs
assigned to worker processes, where either linked lists or some
form of dynamic array lookup is done today.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2696 5f96b256-904b-4d8d-8c98-d829582c6739