core: Log deprecated variables to the 'final' logfile
The previous patch made sure we started writing "real" logmessages
to the right place, but could still cause some logmessages to end up
in the wrong logfile. With this change we protect against that by
adding deprecated variables to a list and printing their warnings
when we're done parsing the entire file (which means we've always
parsed the 'log_file' statement).

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2699 5f96b256-904b-4d8d-8c98-d829582c6739