core: Check timeperiods for circular exclusions
Doing the exclusion logic right means traversing all the exclusions.
If there are circular paths there we might be at it for ever, which
is bad. With this patch we know we don't need to bother checking it
in the timeperiod handling code, which is messy enough as it is.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2700 5f96b256-904b-4d8d-8c98-d829582c6739