cgi: Fix null pointer reference in getcgi.c language selection
Reported as issue 434 to tracker.nagios.org, this only happens on
solaris, where the printf() family functions aren't helpful enough
to replace NULL pointers with the string NULL rather than bombing
out.

While at it, we rename the input parameter so it doesn't collide
with a typename we want to use.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2701 5f96b256-904b-4d8d-8c98-d829582c6739