core: Reorganize and sort-of-simplify base/workers.c
Reorganize because we want the simple file-scoped functions up top,
where pretty much everyone can use them.
Sort-of-simplify because we now have a 1-1 relation (well, almost)
between job creation and destruction, instead of building a job all
over the place and then trying to destroy it in a single function.

This fixes a pretty rare sigsegv problem which could earlier happen
during shutdown if we failed to send one or more jobs to the worker.

The change is larger than it has to be because the bug is extremely
hard to test for, so defensive coding is necessary to avoid triggering
it in the future. It also paves the way for part two of this series,
which will sanitize the worker process management somewhat, since
that's currently somewhat broken for specialized workers.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2718 5f96b256-904b-4d8d-8c98-d829582c6739