Remove unused function parameters
Some functions take parameters that are never used, or are only
used for passing to leaf functions which don't use them. This
removes most of the ones I could easily find, but only where it's
safe to do so with regards to API and ABI compatibility, so the
xdata stuff (which isn't publicly exported to modules) and a lot
of initialization and cleanup functions (which modules aren't
supposed to run anyway) are modified. And CGI's, ofcourse.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2727 5f96b256-904b-4d8d-8c98-d829582c6739