core: Avoid lock file removal race condition
The race condition happens when we receive a SIGHUP, break out
of the event execution loop, and receive any sort of signal
setting the 'sigshutdown' flag AFTER checking for the sigshutdown
state but BEFORE the do-while loop terminates. I.e, during
execution of cleanup() or close_debug_log() towards the end of
the loop.

It's a pretty narrow window, but the result is a segfault on
invalid memory access when we attempt to unlink() the lockfile.

Original-patch-by: Anton Löfgren <alofgren@op5.com>
Signed-off-by: Anton Löfgren <alofgren@op5.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2745 5f96b256-904b-4d8d-8c98-d829582c6739