lib/runcmd: Handle escaped glyphs in double-quoted strings properly
Previously, an argument such as "\\Key\Lalafoo\\Value" would get
translated to "\KeyLalafoo\Value", which is not how the shell does
it.

With this patch we get the properly escaped string \Key\Lalafoo\Value,
which is the same as /bin/echo produces with the above quoting, and
is what Nagios 3 used to execute when given such a string.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2748 5f96b256-904b-4d8d-8c98-d829582c6739