core: Avoid running checks when they're globally disabled
Previously we would execute implicit host dependency checks (for
child and parent hosts) when we received a passive host check
result for either the parent or the child, and and the same went
for both types of host and service dependencies, even though the
check type in question was globally disabled (either by config or
by external command).

This patch alters the behaviour so that disabling checks really
disables them (which I assume is desired), but also means that
submitting passive service checks to hosts and services that in
some way depend upon or are depended upon by other objects can
result in the wrong type of notification being sent out (ie,
"DOWN" instead of "UNREACHABLE" for hosts).

There's nothing we can do about that if we really want to follow
the users wishes, but it's quite a lot easier to explain why
dependency checks aren't working properly when checks are
globally disabled than it is to explain why checks are being
executed.

Signed-off-by: Andreas Ericsson <ae@op5.se>