core: Fix macro initialization during reloads
When we originally start, the statically defined macro_keys array
is initialized with all zeroes, like the rest of the .bss segment,
but when we get a reload signal, it appears it's sometimes changed,
which caused us to url-encode certain macros that really shouldn't
be url-encoded.

With this patch, we should be able to lay issue 407 to rest for
Nagios 4.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2751 5f96b256-904b-4d8d-8c98-d829582c6739