core: free() more variables when quitting
While we'd like to get "No leaks possible" from valgrind, we'll
have to settle for the memory my_strtok() allocates and keeps a
pointer to (for now anyways).

We also make sure to assign more of the "runtime constant" macros
from our global variables, to avoid tracking both of them for leaks.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2762 5f96b256-904b-4d8d-8c98-d829582c6739