xod: Launch a massive strike in the war against strdup()
We already get an allocated and duplicated buffer from mmap_fgets(),
so there's no need for us copy it again, and then once more to get
the value part of the variable in a separate string.

Instead, we avoid the two extra copies per object config entry and
simply modify the input in-place.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2763 5f96b256-904b-4d8d-8c98-d829582c6739