xodtemplate: Don't use space as a delimiter in escalations
contact_names are allowed to have spaces in them, so using space as a
delimiter doesn't really work. Space is thus not a valid delimiter for
contacts anywhere, except in escalations, where it is only valid after
the first comma. So you could have one contact with a space in its name,
as long as you put it first - and no more than that!

This patch allows you to have as many contacts with as many spaces as
you want in your escalations.

This also makes things consistent with how we parse contactgroups for
hostescalations. That particular path did it right all along.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>