core: Don't free the lock file path too soon
It's used after the main loop ends in the main function to remove the
lock file once we're shut down.

As the variable was already freed and pointed to NULL, unlink would not
remove the lock file, but instead return EFAULT. This is not how daemons
are supposed to behave.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>