core: Add perfect-hash sources to version control
It's stupid and useless to have them sitting on my laptop, and it's
even dumber to not use gperf output as-is. This patch fixes both
those issues, allowing anyone to easily extend the types and names
of variables we can get from our worker-processes.

We do check in the output of the gperf run though, as we'd otherwise
introduce a new dependency into the build. Only people who want to
add variables to the worker processes need to bother with the new
dependency.

Signed-off-by: Andreas Ericsson <ae@op5.se>