core: Stop segfaulting when core workers disconnect
Previously we tried to remove the worker from the correct worker
list, but the worker list was a NULL pointer since we never set
the core worker list for core workers.

Even though it's unseemly to brag, I'm kinda proud of the fact that
the workers don't die without being forcefully bludgeoned to death
by a signal.

Signed-off-by: Andreas Ericsson <ae@op5.se>