core: Restore the "Checked %u <objects>" lines for escalations
Apparently, some people use scripts to verify and report how many of
each object type is loaded, and when we removed the "Checked" lines
for host and service escalations (since they're pre-verified), we
screwed those people over. Since it's an easy fix (that also lets us
remove some of the spammier output), we might as well add it back in.

Signed-off-by: Andreas Ericsson <ae@op5.se>