core wproc: Log properly when important helper scripts fail
When important helper-scripts fail, we used to silently ignore that
unless they timed out. Since it's sort of important that eventhandlers
and notification scripts run like they're supposed to, we must log
properly when they don't so post-mortem debugging of missed
notifications and failed eventhandler runs is not only possible, but
positively convenient.

This patch adds a LOT of logging for all non-check system commands
executed that either get killed by a signal, fail to execute
properly or that simply time out.

This means all kinds of helper scripts now use the same API for
logging their state, so a notification script that fails to find
the "sendmail" binary (or whatever) can simply say so on stderr
(or stdout) and then exit with a non-zero exit-status.

Neato.

Signed-off-by: Andreas Ericsson <ae@op5.se>