Stop cgi-bin/status.c from listing unauthorized hosts and services in servicegroup view
Several servicegroup views (overview, summary, grid) in cgi-bin/status.c
list all hosts and services within a servicegroup. This is a security
issue, as hosts and services (at least their names) are leaked to
unauthorized users. Instead, the lists of hosts and services must contain
only objects that the user is authorized to see.

This patch fixes the servicegroup overview, summary and grid views to
list only hosts and services that the user is authorized to see.

Signed-off-by: Andreas Ericsson <ae@op5.se>