xrddefault: Fix loading of notified_on from retention data
There were several problems with how notified_on was loaded from
retention data.

To begin with, notified_on_critical would overwrite the work any other
notified_on* flag had done. That's not very social of it, is it?

To continue, notified_on_warning and notified_on_critical would try to
set their respective bit in the bitmask only if they were *not* set in
the retention data. That's not very logical of it, is it?

Finally, the wrong variable was used for shifting the state, so rather
than setting the flag (1 << STATE), all the fields for both hosts and
services would set the bit at ((1 << STATE) << STATE), leading to values
much larger than expected. That's not very clever of it, is it?

All of these things should be fixed now, so recoveries should once again
work after nagios restarts.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>