xdata/xodtemplate: Fix processing of servicelists for serviceextinfo
This patch fixes two issues with service lookups for serviceextinfo
objects.

Firstly, we fix parameters to xodtemplate_create_service_list(),
which were previously warped so that we shipped the hostgroup_name
as the host_name and vice versa.

Secondly, we reset the master_service_list and service_map for each
iteration to make sure the serviceextinfo isn't set for more objects
than it should be. This second part only matters if there's more
than one serviceextinfo object (which is usually the case if one
uses such deprecated constructs).

Even though serviceextinfo objects are deprecated, we should still
handle them properly if we encounter them.

Original-patch-by: Ricardo Maraschini <ricardo.maraschini@gmail.com>
Commit-message-by: Andreas Ericsson <ae@op5.se>
Signed-off-by: Ricardo Maraschini <ricardo.maraschini@gmail.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>