xdata/xodtemplate: Clear the service bitmap between each serviceextinfo
Without this patch it would be impossible to assign multiple extinfo
objects to the same service, which was possible (although never
officially supported, since it wasn't documented) in Nagios 3.x.

Since it's (marginally) useful and the patch is literally a oneliner
that makes the extinfo code consistent with how escalations and
dependencies work, we may as well put it in (and if "That's what she
said", you need to find yourself a somewhat more passionate partner).

Signed-off-by: Andreas Ericsson <ae@op5.se>