xdata/xodtemplate: Fix the "not enough data" check for serviceextinfo
Since serviceextinfo objects need service_description and either
hostgroup_name or host_name, we should check for just that when
determining if we have enough data or not.

I'm not exactly sure what that check is doing there, since we will
fail with a hard error a few lines later if we can't find the extinfo
targets anyway, but I guess there's some perverted sense of continuity
in perpetuating insane legacies. Besides, people will probably whine
if their configs suddenly stop working, and I really don't want to
deal with that particular grief.

Signed-off-by: Andreas Ericsson <ae@op5.se>