lib/worker: Close input filedescriptors more aggressively
Previously it was possible to end up in the infinite loop of doom
between check_completion() and gather_output() (since the former
calls the latter and vice versa), leading to a coredump when we run
out of stack space. With this patch, that really shouldn't happen,
since we'll now close the offending filedescriptor before we call
out to check_completion() from gather_output(), which means we'll
at most be called upon once to get the final output from each
filedescriptor.

We also avoid calling wait4() without WNOHANG, as that could otherwise
lead us to wait forever in case plugins end up in uninterruptible sleep.

Signed-off-by: Andreas Ericsson <ae@op5.se>