lib/worker: Maintain a kill-queue of processes we fail to reap
Previously it was possible for us to send SIGKILL to our children
and then attempt to reap the child before it had received the signal,
which caused the process to become a zombie.

If we indeed do fail to reap the process after killing it, we'll try
again a few seconds later, and keep on trying every few seconds
until we've either successfully reaped the child, it goes away by
some other means (such as being traced and having its parent
altered), or we get shut down, in which case it's init's problem.

This covers scenarios where the child is stuck in uninterruptable
sleep (such as after a vfork() call, f.e.), but not those where the
plugin itself alters its own process group after its inception.

Signed-off-by: Andreas Ericsson <ae@op5.se>