Avoid crashing when freeing objects after parsing faulty configuration
When we get bad serviceescalations or other entries where we had
already counted the host and servicedependencies but not sorted
them into their list yet, we would go ahead and shut down, calling
the free_object_data() function which would happily try to access
entries in hostdependency_ary and servicedependency_ary, which
caused the crashes mentioned in the subject line.

This patch fixes that particular corner case.

Signed-off-by: Andreas Ericsson <ae@op5.se>