Macros: Make each macro tell which escaping capabilities it accepts
Not all macros should appearently accept command line escaping, but
just a few related to plugin output, and user input.

This commit builds that table again, which was available earlier in
another form. But this time, it isn't used to tell how to escape a
macro, but just which kinds of escaping that should be possible.

While at it, rearrange the macro_key_code struct somewhat to avoid
fill-spill on 64-bit architectures (AE).

Signed-off-by: Andreas Ericsson <ae@op5.se>
Signed-off-by: Max Sikström <msikstrom@op5.com>