lib/worker: Revamp reap timing somewhat for stale jobs
It's easy to artificially create a scenario where our immediate child
goes into uninterruptible sleep with some simple but clearly bananas
programming. Our immediate child, hereafter called "helper", must issue
the vfork() system call and its child may spawn a ton of offspring of
its own if it so wishes, and then go into sleep. If the child's
offspring (the helper's grandchildren) exit, they will become zombies
since the child doesn't wait() for them. The helper on the other hand
is stuck in uninterruptible sleep until its child has called either
exit() or execve(). If it instead calls sleep() the helper can get
stuck in uninterruptible sleep (ps state 'D') for quite some time.

The first time we attempt to reap such a job, the child and all the
childs' children will get the SIGKILL signal and, (unless they too
are stuck in uninterruptible sleep), exit promptly, but not get
wait()ed for, turning them into zombies. The helper on the other hand
will either get its signal queued for when it wakes up (on Linux) or
get its signal discarded (possibly other unices).

The helper will wake up when its immediate child dies, so apparently
that counts as "calling exit", but it will normally not have time to
wake up and smell the signal before the CPU is handed back to the
signalling process (ie, our worker). In such cases the helper will
be ready for reaping very, very soon, but not immediately. When that's
the case, we schedule a 0.25 second sleep before the first attempt at
reaping the check, and then try again every 5 seconds until we've
managed it.

In case the signal is lost instead of queued, that means we'll at
least have killed the helper 0.25 seconds after we get the code back,
and the worst case scenario is that the signal doesn't register until
after we go past the state where we check if the child has exited. In
such cases, we'll leave zombies hanging around for 5 seconds.

The scenario really should be very rare though, and especially on linux
where signals are delivered atomically.

Signed-off-by: Andreas Ericsson <ae@op5.se>