lib/worker: check for completion before we send SIGKILL
Sometimes helpers complete their work just before we attempt to reap
them, in which case the SIGKILL goes to a zombie and we're missing out
on the last chance to reap the actual result of the job instead.

This patch corrects that by forcing one last attempt at properly
reaping the child in case we're about to kill it.

Signed-off-by: Andreas Ericsson <ae@op5.se>