lib/worker: Catch SIGCHLD and schedule a "general" reaping when we do
This means we need a quick way to access the job we just reaped, so
we also add and maintain a fanout table for that precise purpose.

Since we don't expect to run more than a few hundred jobs at the same
time (at worst, that is), we use a relatively small one of 4096 entries
which seems to suffice just fine for our needs.

This appears to eliminate the "checks sometimes reach timeout without
being reaped" more or less completely, unless those checks actually
*do* reach their timeout.

Signed-off-by: Andreas Ericsson <ae@op5.se>